assignment/RAG_PoC/requirements.txt
# Core LangChain
langchain
langchain-community
langchain-openai
langchain-chroma

# Embeddings
openai
sentence-transformers

# Vector Store
chromadb

# Document Loaders
pypdf
unstructured

# Utilities
python-dotenv
tiktoken
